# Conflicts:
#	db/PriceDB.cpp

+++ network/ZMQNetwork.h
@@ -46,7 +46,7 @@ class ZMQNetwork : public TransportNetwork {
 public:
 
 
-    int interruptableRecv(void *_socket, void *_buf, size_t _len, int _flags);
+    uint64_t interruptableRecv(void *_socket, void *_buf, size_t _len, int _flags);
 
     bool interruptableSend(void *_socket, void *_buf, size_t _len, bool _isNonBlocking = false);
